hermes/requirements.txt
browser-use>=0.3.0
langchain-openai>=0.1.0
pyyaml>=6.0
python-telegram-bot>=21.0
httpx>=0.27.0
